[Patch] Set an abstraction to the input system
